requirements.txt
requests
praw
pandas
matplotlib
python-dotenv